backend/requirements.txt
fastapi
uvicorn
python-dotenv
google-genai


httpx
requests
pydantic
scikit-learn
pandas
numpy
google-search-results
